+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [Unreleased]
+
+### Fixed
+
+- Fix handling of more invalid performance starting prices (from @justjasongreen)
+
+
 ## [1.0.0b7] - 2016-07-27
 
 ### Changed
@@ -121,6 +128,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Implement DevOps support (from @justjasongreen)
 
 
+[Unreleased]: https://github.com/justjasongreen/punters_client/compare/1.0.0b7...HEAD
 [1.0.0b7]: https://github.com/justjasongreen/punters_client/compare/1.0.0b6...1.0.0b7
 [1.0.0b6]: https://github.com/justjasongreen/punters_client/compare/1.0.0b5...1.0.0b6
 [1.0.0b5]: https://github.com/justjasongreen/punters_client/compare/1.0.0b4...1.0.0b5